Add a new "Data consumers" team
We are going to consolidate the EFSA & AOPU users into one "can export" team.

First of all, create a new team these users will eventually belong to and give
the team the same permissions as existing ESFA & AOPU teams.

We are adding the locale here because in the Service support area, the teams
are automatically output in the add/edit users screen, and we need a locale
string to ensure the view works correctly.