(Refactor) Change the way feature tests find a particular row to test
In the previous commit, we moved an item in the task list to a new position.
This change meant we had to update many of the feature tests which were looking
for a `li` in `nth-of-type` position, whose position had now changed.

These tests seem quite brittle/unintuitive as they are looking for elements
based on their position in the DOM, as opposed to a more semantic way of finding
elements based on what they contain/are doing on a page.

Update these tests to look for `li` elements based on their content, not their
position in a list.