Fix unknown task identifiers
When loading a task into the edit view it was previously possible to put
in any string as the path and raise an error as the controller tries to
load the class with the string.

This work adds an error when the path cannot be turned into a Task class
and `rescues_from` from the error with a 404 page.