Update by significant date scope
Generally we want to convert anything to an appropriate object as it
enters the system. In this instance the date parameters should be Date
objects unless we have a good reason for them not to be.

We update the scope itself to expect Date objects which allows us to use
a ranged where query rather than a SQL snippet. The new query results in
a `BETWEEN` SQL query which is possibly a better fit.

The all, by month controllers could use further refactoring as they are
working too hard here, but that is not part of this work which focuses
on ensuring Date objects are created as early as possible.