Update by advisory board date scope
Just like the by significant date scope we really want to deal with Date
objects and let Rails create the SQL to actually fetch them.

We want the data entering the system to be turned into appropriate
objects as early as possible, in this case there was no reason to rely
on string at all as the form objects have already cast the parameters to
Dates.