(Fix) Remove strict numericality validation on pupil capacity
A user noticed they were no longer able to add pupil capacity to their task.
After investigation, it seemes that we had `string` db columns for these
values, which should be numbers. Previously, it seems the `only_numeric`
validator would parse strings into numbers for validation, but possibly after
upgrading to Rails 7.1 this no longer works (?)

We are faced with a choice of changing the db columns to integer, or changing
the validation to only use `numericality`, and not `only_numeric`. For now
this is the choice we're making.