Fix the proposed capacity task
Whilst the implementation of this task is not ideal, we know it was
working at some point as we have data in production. Around the time of
the upgrade to Rails 7.1 we had a report of not being able to submit
valid values.

When we investigated we noticed that the columns in the database are
stored as strings, there wasn't any test coverage for the custom data in
the task and the validation was not working as expected.

This work fixes the validation and adds test coverage, however we
decided not to change the database column at this time because we want
to unblock users quickly.