Update date range picker
Our recent refactor of the controller that uses this date range picker
stopped the picker showing the current dates, this was due to how the
pickers were being rendered in the view.

Here we switch to use the more dynamic helpers to render the pickers
(the old ones were hard coded to 2022-2026), showing +/- 2 years of the
current year.

Because we are now relying on Date objects more we have to handle month
numbers below 10 - this is done by casting them to integers just before
we use them.