Refactor all by month transfers controller
This work tries to remove some of the unnecessary redirects in this
controller and to use Date objects where we can, it is a repeat of the
same thing on the equivalent conversions controller.

There is more that can be done with this, but we are keeping this light
touch.