Add confirm headteacher contact details to export
The headteacher needs to be included in the RPA, SUG and FA letters
export.

The role must always be 'Headteacher' so we update the presenter to make
sure this is the case (we have taken steps to ensure that newly added
contacts have the correct role, but this covers legacy contacts as
well).