Marking tasks not applicable clears other values
We have some tasks that when marked as not applicable, clear their other
values. We want this to be the default behaviour across all tasks, which this
work does.

We define a callback for the `save` method on `BaseTaskForm`. With that
in place we can add a callback to the `BaseOptionalTaskForm` which
checks if the task is not applicable and if so, sets all the other
attributes to nil before saving.