Remove logic from primary outgoing trust contact
Having this logic is confusing users, if the primary outgoing trust
contact is not set, showing nothing is a better outcome. We also added
the primary outgoing trust contact role.

The column header is also causing confusion, we are going to use the
phrase 'Primary contact for outgoing trust' for this export as it
matches up with the application UI.