Add basic create transfer service
We already have the conversion version of this service to create new
projects from an API call. This work creates a base class and adds the
transfer projects but does not expose the new API.

At this point we are looking to replicate the same behaviour as the
conversion service. The API handles required fields so that the services
are never called without them.

We changed the approach of the spec as the existing one feels focused on
the user behaviour rather than everything we are trying to do, we'll
come back and refactor the conversion one along the same lines later.