Fix: attempt to find or create user
Because the call to `find_or_create_user` happens regardless of the
validity, there is potential for the URN to be invalid and be used to
try and set the team for a new User.

Whilst this is potentially unlikely, this fix is way less brittle, only
attempting to find or create the user if the parameters appear valid.