Support grouping via API
When projects are added manually, the user can supply a 'group
identifier' from the Prepare application to be used to collect projects
together.

This work allows the group identifier to be sent to the API to create a
new project in a group.

We validate the identifier itself and that the incoming trust UKPRN is
the same as other members of the group.