Add urn validation to API create project service
Just like the form objects that create new projects, the ones for the
API need to validate that no other in progress projects exists for the
same URN.

This work adds a new validator class to handle this as we need to know
the class of the service to know which projects to look for and we need
to check for different project states for projects created from the API,
namely `active` and `inactive`.

We went with a new class rather than extending the existing one to
isolate the behaviour (one day they may not be the ability to add
project manually!?).

We already tidied the transfer spec for this, but the conversion spec
could use some organisation, we'll not doing that here.