Fix and refactor establishment API spec helpers
These are returning the error class rather than the actual errors, in
lots of cases this is fine as we only want the error to have a value,
but we should fix them.

These helper methods have grown over time and are a little disorganised,
so we take the opportunity to tidy up.