Update Academies API client
We can't rely on the bulk endpoints to return 404 if we send a single
parameter that cannot be found, instead we should use the single
endpoint which will tell us if a record cannot be found.

When nothing is found on the bulk endpoints, 200 is returned with an
empty array of results, this is probably fine for us because:

- projects with missing urns are very unlikely due to validation (but
  not impossible!)
- the Project model falls back to fetching the establishment or trust
  itself if the bulk prefetch operation returns nothing

This is all a hangover from when the 'bulk' endpoints were much (much)
faster than the single ones so we used them everywhere (v1 of the API).