Validate incoming trust exists
We want to make sure that the incoming trust can be found on the Academies
API when creating a new project from the API.

Ideally, we only make one call to the API and the response is reused
during User creation.

We needed to update the API spec helpers for this work.