Validate advisory board date for API
We want the same validation as the manual project creation which can be
reused here.

We added the presence validation even though it is not strictly needed
as the Grape API definition requires the parameter.

We need to let Rails know to cast the string parameter to a Date object
for the validation to work, we should declare all of the attribute types
  ideally, but we'll do that in another commit.