Optimise API user create
We can improve this by only initializing the new User if needed and
seeing if it saves, raising the error if it does not.

This saves us the call to `update!` and raises the error on what we
care about, whether the User saved or not as the user shouldn't be
invalid as we validate the email address now.