Do not raise when Academies API 404
When an establishment or trust cannot be found on the Academies API we
treat it as a validation error, i.e. the URN or UKPRN is invalid because
we could not find a match.

Instead of raising a validation error, we should return nil and let the
validation fail, this way all validation errors will be returned instead
of only the not found error.