Better validation errors from the API
Right now we smoosh all the validation errors together into a single
string and use that as the Error message.

This work allows the ValidationError to to accept the ActiveRecod::Errors
and store them, using them to display the errors in a structured manner
in the API response.

We kept the spec simple by turning the errors into json and looking for
strings, this feel good as json is the end result anyway.