Add handover check page
We want users to be able to check they are handing over the correct
project.

We add a link from the index that shows the chosen project and a summary
of the details, users can cancel and return to the index.

Later work will add the behaviour to perform the handover.