Add handover form object
To 'handover' a project created on the API we need to add some
attributes that do not exist in the Prepare application and so cannot be
provided, projects added via the API are invalid until this is done.

The attributes differ for conversions and transfers and the actual
assignment to a user is different depending on the team the project is
assigned to.

As the variation is so small, we kept to this single fancy form object
with two validation contexts. Project use single table inheritance, so
we can assign the project type attributes regardless.

The default context is used when the project is assigned to the Regional
casework services team.

The 'assign' context is used when the project is assigned to one of the
regional teams and so requires extra attributes to be set.

This essentially makes this a very small 'stepped' form with only two
steps, which we have done before, we'll send all the attributes to the
second step when we need to.

We ensure the unused attributes are set to nil so we can simply save
whichever Project is passed in.