Allow handovers
There is a lot here, apologies! This is one journey with four paths and
we are handling it all with the fancy `NewHandoverSteppedForm` object so
it all comes at once!

This delivers a basic, working path for:

- handing over a conversion to Regional Casework Services
- handing over a conversion to a user in the Regional teams
- handing over a transfer to Regional Casework Services
- handing over a transfer to a user in the Regional teams

The approach is that we don't want users to be able to return midway
through the steps (if they are on a path with steps) so we favour
rendering views over new paths - that means passing all the form
parameters to the second step and then posting them off - this feels
okay with so few attributes to manage and we do a similar thing for Date
Histories.

The bulk of the work is in the `HandoversController` which has to know
what type of project and render the right form views.

Most of the changes here are new views and the content to go with them.

We extended the `Projectable` controller concern to take care of invalid
UUIDs - this feels neater than the constraints we have in the routes (we
could switch everything else, but not here!).

We add validation that the email address matches up with a `User`.