Remove duplicate navigation item
We think this is from a time when Service Support could not access the
All projects section. There is no good reasons to have two different
export menus and we loos a tiny bit of view logic!

The method in the export policy is not needed as a User can tell us if
it is service_support.