Change API errors
Returning the same validation as the application UI on the API can be a
little misleading, date formats are different for example.

This work changes from the error messages to the error details,
essentially the keys used to identify the validation error, which should
be descriptive enough for the API as it is only accessed by internal
developers.