Enable the repository to use Codespaces
Unfortunately we have had to remove CSRF protection and bypass the
Azure AD login when running the application in Codespaces.

Github Codespaces runs the application from a random hostname. Azure AD
will not allow sign ins from a hostname it does not recognise, and we
cannot add the random hostname to Azure for every Codespace created.

Similarly, CSRF protection is expecting forms to be submitted from
localhost, not a random hostname, so again we cannot have CSRF
protection turned on when running the application from Codespaces.

Neither of these scenarios are ideal but this will not affect the
application when it is running is production mode.