Fix: attempt to fix flaky autocomplete spec
When running in CI we often see the autocomplete specs fail. We think
this might be caused by the suite running faster than the browser can
render the expected elements.

Elsewhere we have put a short sleep in to give the browser time to
load, so we do the same here.