Allow CI to pass without docker job
The Docker job in our CI fails almost every time due to a know issue.

As we do not require the docker tests to pass in order to pass, we would
rather not show our branches as failing, when they really aren't.

Here we tell Github actions to not fail the CI workflow if the docker job
fails.