Fix: DAO revocation notes on all projects
When we had our first DAO revocation in production it was noted that the
associated note was shown on all projects in the application.

This was caused by the `or` statement not limited the notes to only
those for the project.

A DAO Revocation note should appear in project notes, but only for the
project to which it is associated.

We have removed the `or` altogether in this fix.