Fix autocomplete feature specs
Whilst most of these feature specs work fine, there is one that is
flaky in CI, we see this occasionally, e.g.

https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/actions/runs/11482067947/job/31954234121

Although adding these sleeps slows down the specs, the price is worth it
for consistency.

We basically have to wait for the autocomplete to be rendered so we can
target the resulting elements.