Add length validation to URN and UKPRN parameters in API
We have had several instances of Conversion projects (not
form a MAT) being sent over from Prepare without an incoming
trust UKPRN.

After some investigation we realised the projects were being sent
with a UKRPN of "0", which passes the API validation (which looks
for an integer) but is not a valid UKPRN. So the projects end up
in our "Handover" section with an unknown UKPRN and cannot be
progressed by users.

It seems like Prepare parses a nil UKPRN to 0. If Prepare was
passing nil or an empty string, our validation would have worked.

To stop this happening, tighten up the validation on our side to
look for an integer which is 8 digits long (for UKPRN) or 6 digits
long (for URN) using a regular expression.