Fix: no longer unassign project on update when team is RCS
Previously when editing a project we were unassigning the assigned user
where the project is assigned to the regional casework team. This is a mistake
as assignment to team and assignment to individual are handled separately.

In this commit we separate the handling of assignment on project update:

1. THE ASSIGNED TEAM

This is determined by the answer to:

> Are you handing this project over to RCS (Regional Casework Services)?

and is set with `Project#team`

2. THE ASSIGNED INDIVIDUAL

This may be a caseworker or a regional delivery officer. The assigned user
is changed at:

`/projects/{projectId}/internal-contacts/assigned-user/edit`

and recorded at `Project#assigned_to_id`.