(Fix) Replace OpenStruct with a plain Struct to solve loading issue
On the dev environment (but not locally) we saw an issue with OpenStruct
not being loaded correctly when adding or editing a user in the Service
Support view:

```
ActionView::Template::Error (uninitialized constant User::OpenStruct):
```

We had previously seen a similar issue with the CookiesForm which was
also not reproducible locally, and was fixed in PR #1735

The issue seems to be with OpenStruct not loading or being overriden
by _something_ else, but as it is not reproducible locally it is hard
to diagnose.

In PR #1735 we "fixed" the issue by creating an array of OpenStruct
objects for a 2-answer form. For the create/edit user form, we have
more than 2 answers and we also want these to be drawn from the
`User.user_teams` enum; we don't want to have to manually create an
OpenStruct for each User team in the User model, and create code
duplication.

Instead, use a plain Struct to generate these options from the
User.user_types enum. This does not have the same loading issues as
using OpenStruct.

A Struct is possibly a little slower than using OpenStruct, but as this
form is very rarely used (it is only for Service support users to add
or edit new platform users) this is an acceptable risk.

Also, while this might feel like a regression or a patch as opposed
to a proper fix, we should bear in mind that this application is due to
be replaced soon with a new .Net version.