Play past OpenStruct loading issue with global 'require'
We've been seeing "uninitialize constant" errors for OpenStruct
like

```
uninitialized constant ProjectHelper::OpenStruct
```

which we only see the deployed environments, not in local development
or in CI.

This appears to be a file/class/module loading problem which occurs after
unrelated changes to the codebase.

To play past this issue we now explictly `require "ostruct"` from the
main `application.rb` file.