Check DB connection is available by fetching a record
In a previous commit we attempted to check the DB is available
by using `AxctiveRecord::Base.connection?`. However this will return
false unless you have already opened a connection by (for example)
fetching a record.

Change the healthcheck endpoints to check the db is connected
by both fetching a record and then checking for DB availability.
Put this into a helper module so we can reuse it in both checks.