Enforce uniqueness of URN for active projects
We add a composite uniqueness index on the `projects` table
to ensure that only one project in an active state with a given
URN can exist. Index name:

```
index_projects_on_urn_and_state
```

We've been seeing some duplicate transfers and conversions being
created by our automated "handover" process between "prepare" and
"complete". Our "complete" API receives POSTs to:

- `/api/v1/projects/conversions` and
- `/api/v1/projects/conversions`

these are handled by:

- `Api::Conversions::CreateProjectService` and
- `Api::Transfers::CreateProjectService`

which both use the `UrnUniqueForApiValidator` to perform this
validation at the application level, but for a reliable
constraint we need to use a control at the database layer.

As new projects are added with the "active" state it's sufficient to
use the rule that no two projects may have the same state and URN.

When deploying this change we'll need to identify any existing projects which violate this rule and handle them, probably by moving
all but the most recent duplicate into the deleted state. See the list of [project states][].

[project_states]:
https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/blob/fa42c6bca30696c4ca34b3fe2a3db568244f3294/app/models/project.rb#L92C1-L98C4