Add exception-notification and slack-notifier gems
Although we are currently sending information about unhandled exceptions
through to AppInsights:

`ApplicationInsights::UnhandledException.collect(key)`

these are arriving without an "Operation ID" which our Azure alerting
code expects to find in the `exception` entity, and uses to connect
up with the corresponding `request` entity. Because the exception's
operation ID is blank, no alerting is currently coming through to
Slack.

As:

- it's not obvious how to fix the [Ruby SDK][]
- the Ruby SDK has been "unsupported" for 2 years (and had no
  updates in 6 years)
- this Ruby version of DfE Complete is currently being rewritten in .NET

it's pragmatic to implement the alerting to Slack
within the application. To do that we use the `exception-notification` gem
which has a Slack "notifier". That notifier requires the `slack-notifier`
gem.

[Ruby SDK]:
https://github.com/microsoft/ApplicationInsights-Ruby