Handle long paths used as attacks
The [HighVoltage gem][] was being used to verify that static pages being
requested were valid, i.e. to answer the question

"is `/pages/bananas` a valid static page?"

If the answer was NO then High Voltage was raising an error which
routed to render the not_found template with a 404 status.

However, when passed a very long bogus page ID High Voltage was
throwing an Errno::ENAMETOOLONG error. Which was unhandled.

As HighVoltage is:

- no longer in active development,
- being used soley to verify that there is a static page which matches
  the given `params[:id]`, and as
- there are only 7 static pages

we remove the gem and handle the static pages manually. If the given page
ID isn't recognised our final "mopping up" route renders the `not_found`
404 page:

`match "*unmatched", to: "application#not_found_error", via: :all`

Less magic, more clarity.

[HighVoltage gem]:
https://github.com/thoughtbot/high_voltage