Handle path causing encoding error
We've been seeing errors from requests for paths like

GET '/%ff'

which are causing an InvalidParameterError in Rack which is
captured by Rails as a ActionController::BadRequest.

It's not clear how to get in and rescue this, so in this commit
we create a middleware which reproduces the error and sets the
destination to the root path `/`.