Add note where missing from SignificantDateHistoryReason
We've observed errors in production where a user is viewing the
"Conversion data history" of a project at path such as:

`/projects/{projectId}/date-history`

due to a SignificantDateHistoryReason missing its Note.

On investigation we've found just one Project (and `SignificantDateHistory`
record) affected by this problem where the "reason" for the
change is blank:

```
irb(main):006> SignificantDateHistoryReason.count
=> 3621

irb(main):007> SignificantDateHistoryReason.joins(:note).count
=> 3620
```

The affected record was created on July 1st 2024. We note that
since commit [31b50fb7174651649b385fe683bd384c81e5405f][] June 24th
(Release 75) there has been validation on the presence of the note
to accompany the reason for the change. i.e. the user has to supply
some narrative. See [new_earlier_form_spec.rb][].

Given that it was created after the validation was implemented it's
not understood how this `SignificantDateHistoryReason` comes to be
in this state:

```
{"id"=>"45829896-F491-46D3-9B49-191CFBB7193B",
 "created_at"=>Mon, 01 Jul 2024 10:21:16.484692000 BST +01:00,
 "updated_at"=>Mon, 01 Jul 2024 10:21:16.484692000 BST +01:00,
 "reason_type"=>"buildings",
 "significant_date_history_id"=>"AC104E94-B80F-4F95-B6F0-A7C56DE5A39D"}

irb(main):023> reason.note
=> nil
```

However, for now we take a pragmatic approach:

- trust that all future significant date changes will have the expected
 note(s) attached to their reason(s). Our unit tests verify this.

- create a data migration to identify any `SignificantDateHistoryReason`s
which are missing their `Note` and create the necessary note, with
the content of "Blank (reason unknown)".

That data migration, in this commit, produces output to the logs e.g:

>  Backfilling missing DateHistoryReason notes...
>
>  -> There are 1 reasons which are missing a note:
>
>  Fixing up 1 SignificantDateHistoryReasons missing their Notes:
>
>  * new note created: {
>  	"id"=>"124DDFFE-4504-4C92-B7E5-3863C29C610E",
>	"body"=>"Blank (reason unknown)",
>	"project_id"=>"344F56D4-D52B-4000-A09B-5F94903B8464",
>	"user_id"=>"D84B857D-94D7-4A21-A15D-4904637A21ED",
>	"created_at"=>Tue, 28 Jan 2025 11:20:03.184581000 GMT +00:00,
>	"updated_at"=>Tue, 28 Jan 2025 11:20:03.184581000 GMT +00:00,
>	"task_identifier"=>nil,
>	"notable_id"=>"2BAAA6D5-8340-4342-A2D2-58DDB80D6F4A",
>	"notable_type"=>"SignificantDateHistoryReason"}
>
>  Finished backfilling missing DateHistoryReason notes. There are now 0 reasons missing a note.

[31b50fb7174651649b385fe683bd384c81e5405f]:
https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/commit/31b50fb7174651649b385fe683bd384c81e5405f

[new_earlier_form_spec.rb]:
https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/blob/7026c42ef29b333cccba0aea49e94bc4ec86f103/spec/forms/date_history/reasons/new_earlier_form_spec.rb#L22-L27