Send Ops::ErrorNotification when rescuing API error
Rather than manually triggering an `ExceptionNotification` we
now send an `Ops::ErrorNotification` before showing the user the
`academies_api_client_timeout` template.

This new `Ops::ErrorNotification` will make it clear that this
is a *handled* error. A known problem which we manage gradefully, but on which we want visibility so that we can work on reducing
its frequency.