Add capabilities and user_capabilities tables
At present our handling of user capabilities is handled using roles
defined as attributes on the User entity, which are updated by
a callback which runs `after_save`:

```
private def apply_roles_based_on_team
  assign_attributes(
    assign_to_project: is_regional_caseworker? || is_regional_delivery_officer?,
    manage_user_accounts: apply_service_support_role?,
    manage_conversion_urns: apply_service_support_role?,
    manage_local_authorities: apply_service_support_role?,
    add_new_project: is_regional_delivery_officer?,
    manage_team: apply_team_lead_role?
  )
end
```

In ticket 196438 we have a support request to allow a user access
to two features which are currently mutually exclusive for a
Regional Case Worker: to

- assign unassigned projects
- view their own projects

The definition of "regional case worker" is complex in that draws on
another primary role attribute `manage_team`:

```
def is_regional_caseworker?
  team == "regional_casework_services" && manage_team == false
end

```

In this PR we introduce new concept of `UserCapability` which we
will use to define more granular definitions of what an individual can do.