Show 'find inactive project' when > 1 page of projects to handover
If there is more than one page of projects to handover we now offer a
"Search for inactive project by URN" feature.

If a matching inactive project is found we provide a link to that
project's "check handover details" page. If no matching inactive project
exists we return the user to list of projects to be handed over.

User story
---
- So that I can review a particular project which is to be handed over
- As a Regional Case Officer team leader at a time when there are over
  20 projects in the "to be handed over" (inactive) state
- I want to locate this "to handover" (inactive) project without paging
  through multiple records