Add 'before each' hook to clear all projects & local auths
This is needed in the CI/build as our new trigger is messing with our
test suite's ability to reset the db properly. Perhaps the transaction
used by the trigger to roll back back a transaction which would violate
the URN uniqueness constraint is interfering with the test suite's
transaction rollback?