Link 'By month' to current month's conversions to show active state
Previously the "By month" link in the main navigation was linking to:

`projects/all/by-month/conversions/from/to`
(`date_range_this_month_all_by_month_conversions_projects_path`)

which was showing a list of conversions for the current month. But because
the url was not matching the more literal "Conversions" link in the
sub-nav:

`/projects/all/by-month/conversions/from/3/2025/to/3/2025`
(`date_range_all_by_month_conversions_projects_path`)

The current "Conversions" tab was not being styled as active. In this
commit we change the main nav's "By month" link to use the more explicit
link style for the current month's conversions. In this way we now see
that the default tab is "Conversions".