Fixes post deploy annotation error
When we switched from using workflows to composite actions to build image, a post deploy step 'annotate-release' broke. The step is intended to only run for the 'web' app and not the 'worker', but was attempting to run for both.

We now use v5.1 of DFE-Digital/deploy-azure-container-apps-action/ and set a string  env variable to trigger the annotation only under the 'web' condition.