Describe CachedConnection#fetch(path:, params: nil)
`#fetch` attempts to retrieve and return a cached value. If it's not
found it connects to the Academies API.

It collaborates with the CacheKeyer to get the key for the API call,
based on the path and any params (in the case of bulk requests).

We log either a 'hit' or a 'miss' each time along with the cache key,
path and any params -- to aid debugging.