Update changelog
Defective behaviour
---

Two classes of defective behaviour were reported:

1. Notifications sent when project already assigned

2. Notifications for projects "handed-over" from Prepare not being sent

Changes in this PR
---

When a project is created in Complete by Prepare through the API it
first goes into the "inactive" state. After it's:

- been checked
- had any missing details added
- been successfully "handed over" (which includes having a valid UKPRN
	for the incoming trust)

its state is changed to "active".

If it's assigned to the RCS team, at this point we now send the "new
project to assign" email notification to all RCS team leaders.

We also change the two `EditProjectForm` classes to verify when updating
a project via the UI that the conversion / transfer has not already been
assigned to a caseworker before sending the email notification.