Handover 'Form a MAT' without incoming trust UKPRN
In [PR 2109][] we recently (Release 110 on March 12 2025) added a
guard to the handover controller to show the user an error message if
the project which they were trying to handover was missing its
`incoming_trust_ukprn`. Previously the user was seeing a generic error
page after a validation failure behind the scenes.

It turns out we were a bit too broad with our test, as in the case of
'Form a MAT' projects it's acceptable for the project to have a
`new_trust_reference_number` (TRN) instead of a UKPRN for the
incoming trust. See [3e5cb1be5fe9ee876cbd1e817e113bbdbd9cfd3f][] from
Feb 2024.

In this commit we tweak:

- our handover controller guard to only bail out if *both*
`incoming_trust_ukprn` and `new_trust_reference_number` are blank, and

- the error message to explain that the TRN is an alternative incoming
trust identifier for 'Form a MAT' projects.

[PR 2109]:
https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/pull/2109

[3e5cb1be5fe9ee876cbd1e817e113bbdbd9cfd3f]:
https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/commit/3e5cb1be5fe9ee876cbd1e817e113bbdbd9cfd3f