Describe client's use of CachedConnection
We adjust our tests describing how the client behaves for a range of
successful and unsuccessful requests. Rather than stubbing the Faraday
respond on the client's `connection` we move the stubbing to the
`api_connection` of the client's `cached_connection`.

Not really a unit test but I think that in this case integration with
the collaborating CachedConnection gives us the right level of detail.