Handle Pagy::OverflowError with 404 Not Found
We sometimes see unhandled `Page::OverflowError` exceptions when users
ask our pagination system for a page which doesn't exist, e.g.

```
GET /projects/team/in-progress?page=2
```

Perhaps the user has bookmarked this page or left it open in their
browser overnight. In the interim the number of their
team's "in-progress" projects has reduced, so that they are now listed
on a single page. Asking for page 2  then produces:

```
Page::OverflowError: expected :page in 1..1; got 2
```

We now rescue this error and show the "Not found" template with status
404.