Add Azure management client to fetch rerouting rules
We use a "Managed Identity" based on the AZURE_CLIENT_ID to obtain a
Bearer Token to fetch the Azure Front Door rules from the Azure
Management API.

Note that IDENTITY_ENDPOINT is a special "metadata" endpoint within the
Azure infrastructure. It's mapped to a local port e.g.
`http://localhost42356/msi/token`

See:

- https://learn.microsoft.com/en-us/entra/identity/managed-identities-azure-resources/how-managed-identities-work-vm

and

- https://learn.microsoft.com/en-us/azure/app-service/overview-managed-identity?tabs=portal%2Chttp#rest-endpoint-reference