Update changelog
This service is being replaced by a [new codebase][] written in .NET. In
order to make the transition to the new service gradually this original
version of the service and the new one will run simultaneously, backed
by the same database, with the new service implementing progressively
more elements.

This ["dual running" strategy][] is managed by the Azure FrontDoor CDN which
[is being configured][] to re-reroute specific URL patterns from the
original Ruby service to the new .NET service.

In this PR we implement the user story:

**View effect of rerouting rules**
- So that I can see which of the applications routes are being re-routing
  by the patterns active in Front Door
- As an service support or DevOps user
- I want to see a list of Front Door rerouting patterns and their effect
  on the application's routes

[new codebase][]:
https://github.com/DFE-Digital/complete-conversions-transfers-changes

["dual running" strategy]:
https://dfe-gov-uk.visualstudio.com/Academies-and-Free-Schools-SIP/_wiki/wikis/Academies-and-Free-Schools-SIP.wiki/9064/Rewriting-of-the-Complete-Ruby-product-in-.NET

[is being configured]:
https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/pull/2132